remove more redundant code and the NO_HTMLAGILITYPACK constant
