+++ Libraries/dotNetRDF/Parsing/Contexts/RdfAParserContext.NetCore.cs
@@ -24,7 +24,6 @@
 // </copyright>
 */
 
-#if NO_HTMLAGILITYPACK
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -141,4 +140,3 @@ namespace VDS.RDF.Parsing.Contexts
         }
     }
 }
-#endif
