+++ Libraries/dotNetRDF/Parsing/Contexts/RdfAParserContext.NetFull.cs
@@ -24,7 +24,6 @@
 // </copyright>
 */
 
-#if !NO_HTMLAGILITYPACK
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -140,4 +139,3 @@ namespace VDS.RDF.Parsing.Contexts
         }
     }
 }
-#endif
