+++ Libraries/dotNetRDF/Parsing/RdfAParser.NetFull.cs
@@ -24,24 +24,23 @@
 // </copyright>
 */
 
-#if NO_HTMLAGILITYPACK
-using AngleSharp.Dom;
-using AngleSharp.Parser.Html;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-#if !NO_WEB
 using System.Web;
-#endif
+using HtmlAgilityPack;
 using VDS.RDF.Parsing.Contexts;
 using VDS.RDF.Parsing.Handlers;
 using VDS.RDF.Query;
 
+// SLT: Find alternative to HttpUtility.DecodeHtml for Silverlight
+
 namespace VDS.RDF.Parsing
 {
+
     /// <summary>
     /// Class for reading RDF embedded as RDFa from within HTML web pages
     /// </summary>
@@ -154,8 +153,8 @@ namespace VDS.RDF.Parsing
 
             try
             {
-                var parser = new HtmlParser();
-                var doc = parser.Parse(input.ReadToEnd());
+                HtmlDocument doc = new HtmlDocument();
+                doc.Load(input);
 
                 RdfAParserContext context = new RdfAParserContext(handler, doc);
                 this.Parse(context);
@@ -187,7 +186,7 @@ namespace VDS.RDF.Parsing
         {
             if (handler == null) throw new RdfParseException("Cannot read RDF into a null RDF Handler");
             if (filename == null) throw new RdfParseException("Cannot read RDF from a null File");
-            Load(handler, new StreamReader(File.Open(filename, FileMode.Open), Encoding.UTF8));
+            this.Load(handler, new StreamReader(filename, Encoding.UTF8));
         }
 
         private void Parse(RdfAParserContext context)
@@ -205,14 +204,12 @@ namespace VDS.RDF.Parsing
                 evalContext.LocalVocabulary = new TermMappings();
 
                 // If there's a base element this permanently changes the Base URI
-                var baseEl = context.Document.QuerySelector("head > base");
-                //HtmlNode baseEl = context.Document.DocumentNode.SelectSingleNode("/html/head/base");
+                HtmlNode baseEl = context.Document.DocumentNode.SelectSingleNode("/html/head/base");
                 if (baseEl != null)
                 {
-                    //if (baseEl.Attributes.Contains("href"))
-                    var uri = baseEl.GetAttribute("href");
-                    if (!String.IsNullOrEmpty(uri))
+                    if (baseEl.Attributes.Contains("href"))
                     {
+                        String uri = baseEl.Attributes["href"].Value;
                         if (uri.Contains("?"))
                         {
                             evalContext.BaseUri = UriFactory.Create(uri.Substring(0, uri.IndexOf('?')));
@@ -229,13 +226,25 @@ namespace VDS.RDF.Parsing
                 }
 
                 // Check whether xml:base is permissible
-                var docType = context.Document.Doctype.SystemIdentifier;
-                if (!String.IsNullOrEmpty(docType))
+                HtmlNodeCollection docTypes = context.Document.DocumentNode.SelectNodes("comment()");
+                if (docTypes != null)
                 {
-                    if (docType.Equals(XHtmlPlusRdfADoctype))
+                    foreach (HtmlNode docType in docTypes)
                     {
-                        // XHTML+RDFa does not permit xml:base
-                        context.XmlBaseAllowed = false;
+                        if (docType.InnerText.StartsWith("<!DOCTYPE"))
+                        {
+                            // Extract the Document Type
+                            Match dtd = Regex.Match(docType.InnerText, "\"([^\"]+)\">");
+                            if (dtd.Success)
+                            {
+                                if (dtd.Groups[1].Value.Equals(XHtmlPlusRdfADoctype))
+                                {
+                                    // XHTML+RDFa does not permit xml:base
+                                    context.XmlBaseAllowed = false;
+                                }
+                                break;
+                            }
+                        }
                     }
                 }
 
@@ -243,8 +252,16 @@ namespace VDS.RDF.Parsing
                 context.Syntax = this._syntax;
                 if (context.Syntax == RdfASyntax.AutoDetect || context.Syntax == RdfASyntax.AutoDetectLegacy)
                 {
-                    var docNode = context.Document.QuerySelector("html");
-                    if (docNode != null && docNode.HasAttribute("version"))
+                    HtmlNode docNode;
+                    if (context.Document.DocumentNode.Name.Equals("html"))
+                    {
+                        docNode = context.Document.DocumentNode;
+                    }
+                    else
+                    {
+                        docNode = context.Document.DocumentNode.SelectSingleNode("html");
+                    }
+                    if (docNode != null && docNode.Attributes.Contains("version"))
                     {
                         String version = docNode.Attributes["version"].Value;
                         switch (version)
@@ -288,10 +305,7 @@ namespace VDS.RDF.Parsing
                     }
                 }
 
-                foreach (var child in context.Document.Children)
-                {
-                    this.ProcessElement(context, evalContext, child);
-                }
+                this.ProcessElement(context, evalContext, context.Document.DocumentNode);
 
                 context.Handler.EndRdf(true);
             }
@@ -307,21 +321,21 @@ namespace VDS.RDF.Parsing
             }
         }
 
-        private void ProcessElement(RdfAParserContext context, RdfAEvaluationContext evalContext, IElement currElement)
+        private void ProcessElement(RdfAParserContext context, RdfAEvaluationContext evalContext, HtmlNode currElement)
         {
             bool recurse = true, skip = false;
             bool rel = false,
                  rev = false,
                  about = false,
                  src = false,
-                 href = false,
+                 href = false, 
                  property = false,
                  type = false,
                  resource = false,
                  content = false,
                  datatype = false;
             bool noDefaultNamespace = false;
-
+            
             INode newSubj = null, currObj = null;
             List<IncompleteTriple> incomplete = new List<IncompleteTriple>();
             List<String> inScopePrefixes = new List<string>();
@@ -336,7 +350,7 @@ namespace VDS.RDF.Parsing
             #region Steps 2-5 of the RDFa Processing Rules
 
             // Locate namespaces and other relevant attributes
-            foreach (var attr in currElement.Attributes)
+            foreach (HtmlAttribute attr in currElement.Attributes)
             {
                 String uri;
                 if (attr.Name.StartsWith("xmlns:"))
@@ -455,7 +469,7 @@ namespace VDS.RDF.Parsing
                                 }
                                 else
                                 {
-                                    this.OnWarning("Unable to resolve a Profile document specified by the @profile attribute on the element <" + currElement.LocalName + "> - ignoring the DOM subtree of this element");
+                                    this.OnWarning("Unable to resolve a Profile document specified by the @profile attribute on the element <" + currElement.Name + "> - ignoring the DOM subtree of this element");
                                     return;
                                 }
                             }
@@ -483,7 +497,7 @@ namespace VDS.RDF.Parsing
             // Calls to Tools.ResolveUri which error will still cause the parser to halt
             if (!rel && !rev)
             {
-                // No @rel or @rev attributes
+                 // No @rel or @rev attributes
                 if (about && !currElement.Attributes["about"].Value.Equals("[]"))
                 {
                     // New Subject is the URI
@@ -504,7 +518,7 @@ namespace VDS.RDF.Parsing
                     // New Subject is the URI
                     newSubj = context.Handler.CreateUriNode(UriFactory.Create(Tools.ResolveUri(currElement.Attributes["href"].Value, baseUri)));
                 }
-                else if (currElement.LocalName.Equals("head") || currElement.LocalName.Equals("body"))
+                else if (currElement.Name.Equals("head") || currElement.Name.Equals("body"))
                 {
                     // New Subject is the Base URI
                     try
@@ -551,7 +565,7 @@ namespace VDS.RDF.Parsing
                     // New Subject is the URI
                     newSubj = context.Handler.CreateUriNode(UriFactory.Create(Tools.ResolveUri(currElement.Attributes["src"].Value, baseUri)));
                 }
-                else if (currElement.LocalName.Equals("head") || currElement.LocalName.Equals("body"))
+                else if (currElement.Name.Equals("head") || currElement.Name.Equals("body"))
                 {
                     // New Subject is the Base URI
                     try
@@ -706,7 +720,7 @@ namespace VDS.RDF.Parsing
                         {
                             // Value is concatentation of child text nodes
                             StringBuilder lit = new StringBuilder();
-                            foreach (var n in currElement.ChildNodes)
+                            foreach (HtmlNode n in currElement.ChildNodes)
                             {
                                 this.GrabText(lit, n);
                             }
@@ -717,7 +731,7 @@ namespace VDS.RDF.Parsing
                     {
                         // It's an XML Literal - this is now RDFa 1.0 Only
                         // This is an incompatability with RDFa 1.1
-                        foreach (var child in currElement.ChildNodes)
+                        foreach (HtmlNode child in currElement.ChildNodes)
                         {
                             this.ProcessXmlLiteral(evalContext, child, noDefaultNamespace);
                         }
@@ -730,7 +744,7 @@ namespace VDS.RDF.Parsing
                         dtNode = null;
                     }
                 }
-
+                
                 if (dtNode == null)
                 {
                     // A Plain Literal
@@ -742,17 +756,17 @@ namespace VDS.RDF.Parsing
                     else if (!currElement.HasChildNodes)
                     {
                         // Value is content of the element (if any)
-                        currLiteral = context.Handler.CreateLiteralNode(HttpUtility.HtmlDecode(currElement.TextContent), lang);
+                        currLiteral = context.Handler.CreateLiteralNode(HttpUtility.HtmlDecode(currElement.InnerText), lang);
                     }
-                    else if (currElement.ChildNodes.All(n => n.NodeType == AngleSharp.Dom.NodeType.Text))
+                    else if (currElement.ChildNodes.All(n => n.NodeType == HtmlNodeType.Text))
                     {
                         // Value is concatenation of all Text Child Nodes
                         StringBuilder lit = new StringBuilder();
-                        foreach (var n in currElement.ChildNodes)
+                        foreach (HtmlNode n in currElement.ChildNodes)
                         {
-                            if (n.NodeType == AngleSharp.Dom.NodeType.Text)
+                            if (n.NodeType == HtmlNodeType.Text)
                             {
-                                lit.Append(n.TextContent);
+                                lit.Append(n.InnerText);
                             }
                         }
                         currLiteral = context.Handler.CreateLiteralNode(HttpUtility.HtmlDecode(lit.ToString()), lang);
@@ -760,7 +774,7 @@ namespace VDS.RDF.Parsing
                     else if (!datatype || (datatype && currElement.Attributes["datatype"].Equals(String.Empty)))
                     {
                         // Value is an XML Literal
-                        foreach (var child in currElement.ChildNodes)
+                        foreach (HtmlNode child in currElement.ChildNodes)
                         {
                             this.ProcessXmlLiteral(evalContext, child, noDefaultNamespace);
                         }
@@ -847,11 +861,11 @@ namespace VDS.RDF.Parsing
                     newEvalContext.LocalVocabulary = new TermMappings(evalContext.LocalVocabulary);
 
                     // Iterate over the Nodes
-                    foreach (var n in currElement.ChildNodes)
+                    foreach (HtmlNode n in currElement.ChildNodes)
                     {
-                        if (n.NodeType == AngleSharp.Dom.NodeType.Element)
+                        if (n.NodeType == HtmlNodeType.Element)
                         {
-                            this.ProcessElement(context, newEvalContext, n as IElement);
+                            this.ProcessElement(context, newEvalContext, n);
                         }
                     }
                 }
@@ -1094,8 +1108,8 @@ namespace VDS.RDF.Parsing
             if (value.Contains(" "))
             {
                 values = value.Split(' ');
-            }
-            else
+            } 
+            else 
             {
                 values = new String[] { value };
             }
@@ -1132,8 +1146,8 @@ namespace VDS.RDF.Parsing
             if (value.Contains(" "))
             {
                 values = value.Split(' ');
-            }
-            else
+            } 
+            else 
             {
                 values = new String[] { value };
             }
@@ -1155,7 +1169,7 @@ namespace VDS.RDF.Parsing
             return nodes;
         }
 
-        private void ParsePrefixAttribute(RdfAParserContext context, RdfAEvaluationContext evalContext, IAttr attr, String baseUri, Dictionary<string, Uri> hiddenPrefixes, List<String> inScopePrefixes)
+        private void ParsePrefixAttribute(RdfAParserContext context, RdfAEvaluationContext evalContext, HtmlAttribute attr, String baseUri, Dictionary<string,Uri> hiddenPrefixes, List<String> inScopePrefixes)
         {
             // Do nothing if the @prefix attribute is empty
             if (attr.Value.Equals(String.Empty)) return;
@@ -1164,7 +1178,7 @@ namespace VDS.RDF.Parsing
             char next;
             bool canExit = false;
 
-            do
+            do 
             {
                 StringBuilder prefixData = new StringBuilder();
                 StringBuilder uriData = new StringBuilder();
@@ -1238,7 +1252,7 @@ namespace VDS.RDF.Parsing
             } while (!canExit);
         }
 
-        private bool ParseProfileAttribute(RdfAParserContext context, RdfAEvaluationContext evalContext, IAttr attr)
+        private bool ParseProfileAttribute(RdfAParserContext context, RdfAEvaluationContext evalContext, HtmlAttribute attr)
         {
             String[] profiles;
             if (attr.Value.Contains(" "))
@@ -1259,7 +1273,7 @@ namespace VDS.RDF.Parsing
                 {
                     Graph g = new Graph();
 
-                    if (profile.Equals(XHtmlVocabNamespace) || profile.Equals(XHtmlVocabNamespace.Substring(0, XHtmlVocabNamespace.Length - 1)))
+                    if (profile.Equals(XHtmlVocabNamespace) || profile.Equals(XHtmlVocabNamespace.Substring(0, XHtmlVocabNamespace.Length-1)))
                     {
                         // XHTML Vocabulary is a fixed vocabulary
                         evalContext.LocalVocabulary.Merge(new XHtmlRdfAVocabulary());
@@ -1329,7 +1343,7 @@ namespace VDS.RDF.Parsing
             return true;
         }
 
-        private void ParseVocabAttribute(RdfAParserContext context, RdfAEvaluationContext evalContext, IAttr attr)
+        private void ParseVocabAttribute(RdfAParserContext context, RdfAEvaluationContext evalContext, HtmlAttribute attr)
         {
             if (attr.Value.Equals(String.Empty))
             {
@@ -1342,34 +1356,31 @@ namespace VDS.RDF.Parsing
             }
         }
 
-        private void GrabText(StringBuilder output, AngleSharp.Dom.INode n)
+        private void GrabText(StringBuilder output, HtmlNode n)
         {
-            // TextContent property includes contet of descendants
-            output.Append(n.TextContent);
-            //switch (n.NodeType)
-            //{
-            //    case AngleSharp.Dom.NodeType.Document:
-            //    case AngleSharp.Dom.NodeType.Element:
-            //        foreach (var child in n.ChildNodes)
-            //        {
-            //            this.GrabText(output, child);
-            //        }
-            //        break;
-            //    case AngleSharp.Dom.NodeType.Text:
-            //        output.Append(n.InnerText);
-            //        break;
-            //}
+            switch (n.NodeType)
+            {
+                case HtmlNodeType.Document:
+                case HtmlNodeType.Element:
+                    foreach (HtmlNode child in n.ChildNodes)
+                    {
+                        this.GrabText(output, child);
+                    }
+                    break;
+                case HtmlNodeType.Text:
+                    output.Append(n.InnerText);
+                    break;
+            }
         }
 
-        private void ProcessXmlLiteral(RdfAEvaluationContext evalContext, AngleSharp.Dom.INode n, bool noDefaultNamespace)
+        private void ProcessXmlLiteral(RdfAEvaluationContext evalContext, HtmlNode n, bool noDefaultNamespace)
         {
-            if (n.NodeType == AngleSharp.Dom.NodeType.Element)
+            if (n.NodeType == HtmlNodeType.Element)
             {
-                var element = n as IElement;
                 // Add Default Namespace as XHTML Namespace unless this would override an existing namespace
-                if (!element.HasAttribute("xmlns"))
+                if (!n.Attributes.Contains("xmlns"))
                 {
-                    if (!noDefaultNamespace) element.SetAttribute("xmlns", XHtmlNamespace);
+                    if (!noDefaultNamespace) n.Attributes.Add("xmlns", XHtmlNamespace);
                 }
                 else
                 {
@@ -1377,9 +1388,7 @@ namespace VDS.RDF.Parsing
                 }
 
                 // Add specific namespaces if necessary
-                /*
-                 * KA: Not sure if this section is needed any more, as it looks like AngleSharp handles the namespace mapping
-                if (!String.IsNullOrEmpty(element.NamespaceUri))
+                if (n.Name.Contains(":"))
                 {
                     String prefix = n.Name.Substring(0, n.Name.IndexOf(':'));
                     if (n.Attributes.Contains("xmlns:" + prefix))
@@ -1396,20 +1405,19 @@ namespace VDS.RDF.Parsing
                         throw new RdfParseException("Malformed XML Literal - the undefined namespace prefix '" + prefix + "' is used");
                     }
                 }
-                */
                 // Add Language (but don't override existing language)
                 if (!evalContext.Language.Equals(String.Empty))
                 {
-                    if (!element.HasAttribute("xml:lang"))
+                    if (!n.Attributes.Contains("xml:lang"))
                     {
-                        element.SetAttribute("xml:lang", evalContext.Language);
+                        n.Attributes.Add("xml:lang", evalContext.Language);
                     }
                 }
 
                 // Recurse on any child nodes
-                foreach (var child in n.ChildNodes)
+                foreach (HtmlNode child in n.ChildNodes)
                 {
-                    ProcessXmlLiteral(evalContext, child, noDefaultNamespace);
+                    this.ProcessXmlLiteral(evalContext, child, noDefaultNamespace);
                 }
             }
         }
@@ -1431,8 +1439,8 @@ namespace VDS.RDF.Parsing
                 String prefix = value.Substring(0, value.IndexOf(':'));
                 String reference = value.Substring(value.IndexOf(':') + 1);
                 return (XmlSpecsHelper.IsNCName(prefix) || prefix.Equals("_")) && evalContext.NamespaceMap.HasNamespace(prefix) && IriSpecsHelper.IsIrelativeRef(reference);
-            }
-            else
+            } 
+            else 
             {
                 return false;
             }
@@ -1457,4 +1465,3 @@ namespace VDS.RDF.Parsing
         public event RdfReaderWarning Warning;
     }
 }
-#endif
