+++ Libraries/dotNetRDF/Parsing/StringParser.cs
@@ -112,14 +112,12 @@ namespace VDS.RDF.Parsing
                     format = "RDF/XML";
                     Parse(g, data, new RdfXmlParser());
                 }
-#if !NO_HTMLAGILITYPACK
                 else if (data.Contains("<html"))
                 {
                     // HTML (possibly containing RDFa)
                     format = "HTML+RDFa";
                     Parse(g, data, new RdfAParser());
                 }
-#endif
                 else if (data.Contains("@prefix") || data.Contains("@base"))
                 {
                     // Turtle/Notation 3
@@ -333,13 +331,11 @@ namespace VDS.RDF.Parsing
                 // Probably RDF/XML
                 return new RdfXmlParser();
             }
-#if !NO_HTMLAGILITYPACK
             else if (data.Contains("<html"))
             {
                 // HTML (possibly containing RDFa)
                 return new RdfAParser();
             }
-#endif
             else if (data.Contains("@prefix") || data.Contains("@base"))
             {
                 // Turtle/Notation 3
