+++ Testing/unittest/Parsing/Suites/RdfA.cs
@@ -23,8 +23,6 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#if !NO_HTMLAGILITYPACK
-
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -39,8 +37,6 @@ using VDS.RDF.Writing;
 
 namespace VDS.RDF.Parsing.Suites
 {
-#if !NO_HTMLAGILITYPACK
-
     public class RdfA
     {
         private void reportError(String header, Exception ex)
@@ -285,6 +281,4 @@ namespace VDS.RDF.Parsing.Suites
             Console.WriteLine("Warning: " + warning);
         }
     }
-#endif
 }
-#endif
