+++ appveyor.yml
@@ -28,13 +28,7 @@ build_script:
   - dotnet pack -c Release -o %APPVEYOR_BUILD_FOLDER%\nugets Libraries/dotNetRDF.Query.Spin /p:Version=%GitVersion_NuGetVersion%
 
 test_script:
-  - ps: |
-      dotnet test .\Testing\unittest\unittest.csproj -l "trx;LogFileName=output.trx" --filter category!=explicit --no-build | Out-File .\test.log
-      if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
-      dotnet test .\Testing\unittest\unittest.csproj -l "trx;LogFileName=fulltext.trx" --filter category=fulltext --no-build | Out-File .\fulltext.log
-      if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
-      dotnet test .\Testing\netcoretest\netcoretest.csproj -l "trx;LogFileName=netcore.trx" --no-build | Out-File .\netcoretest.log
-      if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
+  - dotnet xunit .\Testing\unittest\unittest.csproj -configuration %CONFIGURATION% > test.log
 
 after_test:
   - cmd: msbuild Build\shfb\dotnetrdf.shfbproj /logger:"C:\Program Files\AppVeyor\BuildAgent\Appveyor.MSBuildLogger.dll" /p:Configuration=%CONFIGURATION% /p:HelpFileVersion=%GitVersion_NuGetVersion%
@@ -59,20 +53,8 @@ deploy:
 on_finish:
   - ps: |
       $ws = (New-Object 'System.Net.WebClient')
-      if((Test-Path .\Testing\unittest\TestResults\output.trx)) {    
-        Push-AppveyorArtifact .\Testing\unittest\TestResults\output.trx
+      if((Test-Path .\test.log)) {    
         Push-AppveyorArtifact .\test.log
-        $ws.UploadFile("https://ci.appveyor.com/api/testresults/mstest/$($env:APPVEYOR_JOB_ID)", (Resolve-Path .\Testing\unittest\TestResults\output.trx ) )
-      }
-      if((Test-Path .\Testing\unittest\TestResults\fulltext.trx)) {
-        Push-AppveyorArtifact .\Testing\unittest\TestResults\fulltext.trx
-        Push-AppveyorArtifact .\fulltext.log
-        $ws.UploadFile("https://ci.appveyor.com/api/testresults/mstest/$($env:APPVEYOR_JOB_ID)", (Resolve-Path .\Testing\unittest\TestResults\fulltext.trx ) )
-      }
-      if((Test-Path .\Testing\netcoretest\TestResults\netcore.trx)) {
-        Push-AppveyorArtifact .\Testing\netcoretest\TestResults\netcore.trx
-        Push-AppveyorArtifact .\netcoretest.log
-        $ws.UploadFile("https://ci.appveyor.com/api/testresults/mstest/$($env:APPVEYOR_JOB_ID)", (Resolve-Path .\Testing\netcoretest\TestResults\netcore.trx ) )
       }
 
 artifacts:
